dom/media/webaudio/test/test_mediaStreamAudioDestinationNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test MediaStreamAudioDestinationNode</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<audio id="audioelem"></audio>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("This test uses a live media element so it needs to wait for the media stack to do some work.");
addLoadEvent(function() {
  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }

  var source = context.createBufferSource();
  source.buffer = buffer;

  var dest = new MediaStreamAudioDestinationNode(context);
  source.connect(dest);

  var elem = document.getElementById('audioelem');
  elem.srcObject = dest.stream;
  elem.onloadedmetadata = function() {
    ok(true, "got metadata event");
    setTimeout(function() {
      is(elem.played.length, 1, "should have a played interval");
      is(elem.played.start(0), 0, "should have played immediately");
      isnot(elem.played.end(0), 0, "should have played for a non-zero interval");

      // This will end the media element.
      dest.stream.getTracks()[0].stop();
    }, 2000);
  };
  elem.onended = function() {
    ok(true, "media element ended after destination track.stop()");
    SimpleTest.finish();
  };

  source.start(0);
  elem.play();
});
</script>
